requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
executing==2.0.1
fastapi==0.109.1
h11==0.14.0
icecream==2.1.3
idna==3.6
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pymongo==4.6.1
PyMySQL==1.1.0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
